+++ routes/account-routes.tsx
@@ -1,0 +1,37 @@
+import {
+  getLatestRoomNameForOwnerName,
+  getUrlForRoomNameAndOwner,
+} from "../lib/database-access.ts";
+
+"/account";
+import { redirect, wrapReactElem } from "./helpers.tsx";
+import { getFullRoomUrlFromUrlName } from "../lib/utils.ts";
+import { AccountPage } from "../web-server/components.tsx";
+import { Routes } from "../lib/types.ts";
+
+export const account: Routes = {
+  "/account": {
+    GET: ({ requestAuthContext }) => {
+      const user = requestAuthContext.getUser();
+      if (!user) {
+        return redirect("/login");
+      }
+      const state: { username: string; roomUrl?: string; roomName?: string } = {
+        username: user.username,
+      };
+      const latestRoomName = getLatestRoomNameForOwnerName(user.username);
+      if (latestRoomName) {
+        const urlName = getUrlForRoomNameAndOwner(
+          latestRoomName,
+          user.username,
+        );
+        if (urlName) {
+          state.roomName = latestRoomName;
+          state.roomUrl = getFullRoomUrlFromUrlName(urlName);
+        }
+      }
+
+      return wrapReactElem(AccountPage(state), state);
+    },
+  },
+};
